requirements.txt
keras==3.4.1
opencv-python==4.8.0.74
pillow==11.0.0
scikit-learn==1.4.2
seaborn==0.13.2
numpy==1.26.4
tensorflow==2.17.0
keras==3.4.1
matplotlib==3.7.2
setuptools
pip==24.3.1
streamlit
opencv-python-headless
